MesaLib18: adjust pthread stubs patch to ensure no OS picks it up, as we don't provide libpthread stubs.
Note that the name is misleading, libpthread stubs refers to a library
that only pulls in libpthread if a set of functions is not available as
stubs as libc. It is only a pkgconfig file that answers the question
"pull in libpthread? Y/N".

Replace non-existent libpthread stubs buildlink3.mk with the pthread
buildlink3.mk, we should be linking with libpthread for most (all?)
OSes.