xf86-video-r128: update to 6.12.0.
This release is mainly meant deal with 2 urgent build failure issues
reported by users.
If your distribution is using Version 6.11.0, please replace it
with this version as soon as possible.
There is very little change in the functionality compared to the
previous release, but I did rewrite hard to maintain and very
awkwardly written XAA / EXA initialization code.
It is always possible that I can mess up the code, but it appears
that the code is working correctly.